TR-12984 Extract communication to the appcatalog package
