+++ submit.go
@@ -1,23 +1,14 @@
 package appix
 
 import (
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
-	"net/http"
 
 	kingpin "gopkg.in/alecthomas/kingpin.v2"
 
 	"github.com/Travix-International/appix/appcatalog"
-	"github.com/Travix-International/appix/auth"
 	"github.com/Travix-International/appix/config"
 )
-
-type submitResponse struct {
-	Messages []string
-	Links    map[string]string
-}
 
 // RegisterSubmit registers the 'submit' command.
 func RegisterSubmit(app *kingpin.Application, config config.Config, args *GlobalArgs) {
@@ -51,83 +42,22 @@ func RegisterSubmit(app *kingpin.Application, config config.Config, args *Global
 
 			rootURI := config.CatalogURIs[environment]
 			submitURI := fmt.Sprintf(submitTemplateURI, rootURI, appName)
-			files := map[string]string{
-				"manifest": appManifestFile,
-				"zapfile":  zapFile,
-			}
 
-			if args.Verbose {
-				log.Println("Posting files to App Catalog: " + submitURI)
-			}
-			request, err := appcatalog.CreateMultiFileUploadRequest(submitURI, files, nil, args.Verbose)
+			acceptanceQueryURLPath, err := appcatalog.SubmitToCatalog(submitURI, appManifestFile, zapFile, args.Verbose, config)
+
 			if err != nil {
-				log.Println("Call to App Catalog failed!")
 				return err
 			}
 
-			token, err := auth.LoadAuthToken(config)
+			log.Println("App has been submitted successfully.")
 
-			if err == nil {
-				request.Header.Set("Authorization", token.TokenType+" "+token.IdToken)
-			} else {
-				log.Println("WARNING: You are not logged in. In a future version authentication will be mandatory.\nYou can log in using \"appix login\".")
-			}
-
-			client := &http.Client{}
-			response, err := client.Do(request)
-			if err != nil {
-				log.Println("Call to App Catalog failed!")
-				return err
-			}
-
-			if response.StatusCode == 401 || response.StatusCode == 403 {
-				log.Printf("You are not authorized to submit the application to the App Catalog (status code %v). If you are not signed in, please log in using 'appix login'.", response.StatusCode)
-				return fmt.Errorf("Authentication error")
-			}
-
-			responseBody, err := ioutil.ReadAll(response.Body)
-			if err != nil {
-				log.Println("Error reading response from App Catalog!")
-				return err
-			}
-
-			var responseObject submitResponse
-			err = json.Unmarshal(responseBody, &responseObject)
-			if err != nil {
-				if args.Verbose {
-					log.Println(err)
-				}
-
-				responseObject = submitResponse{}
-				responseObject.Messages = []string{string(responseBody)}
-			}
-
-			log.Printf("App Catalog returned statuscode %v. Response details:\n", response.StatusCode)
-			for _, line := range responseObject.Messages {
-				log.Printf("\t%v\n", line)
-			}
-
-			if args.Verbose {
-				for key, val := range responseObject.Links {
-					log.Printf("\tLINK: %s\t\t%s", key, val)
-				}
-			}
-
-			if response.StatusCode == http.StatusOK {
-				log.Println("App has been submitted successfully.")
-
-				if acceptanceQueryUrlPath, ok := responseObject.Links["acc:query"]; ok {
-					log.Println("You can use the following query URL to get this particular version of this app:")
-					log.Printf("\t%s%s\n", rootURI, acceptanceQueryUrlPath)
-				}
-			} else {
-				return fmt.Errorf("Submit failed, App Catalog returned statuscode %v", response.StatusCode)
+			if acceptanceQueryURLPath != "" {
+				log.Println("You can use the following query URL to get this particular version of this app:")
+				log.Printf("\t%s%s\n", rootURI, acceptanceQueryURLPath)
 			}
 
 			return nil
-		}).
-		Alias("pub").
-		Alias("publ")
+		})
 
 	command.Arg("appPath", "path to the App folder (default: current folder)").
 		Default(".").
